Restore tests for buildFromApiRequest now testing against DonationService
