+++ app/repositories.php
@@ -57,7 +57,6 @@ return static function (ContainerBuilder $containerBuilder) {
 
             \assert($repo instanceof DoctrineDonationRepository);
 
-            $repo->setCampaignRepository($c->get(CampaignRepository::class));
             $repo->setClient($c->get(Client\Donation::class));
             $repo->setLogger($c->get(LoggerInterface::class));
             $repo->setMatchingAdapter($c->get(Matching\Adapter::class));
