+++ src/Domain/DoctrineDonationRepository.php
@@ -33,8 +33,6 @@ class DoctrineDonationRepository extends SalesforceProxyRepository implements Do
 
     private const int MAX_SALEFORCE_FIELD_UPDATE_TRIES = 3;
 
-    private CampaignRepository $campaignRepository;
-    private FundRepository $fundRepository;
     /**
      * If changing the value of EXPIRY_SECONDS make sure to update environment.reservationMinutes to match in
      * donate-frontend (or consider making frontend use expiration dates generated by matchbot)
@@ -50,11 +48,12 @@ class DoctrineDonationRepository extends SalesforceProxyRepository implements Do
         $this->matchingAdapter = $adapter;
     }
 
-    /**
-     * @deprecated - use {@see DonationService::buildFromAPIRequest() directly instead}
-     */
-    public function buildFromApiRequest(DonationCreate $donationData, PersonId $donorId, DonationService $donationService): Donation
-    {
+    #[\Override]
+    public function buildFromApiRequest(
+        DonationCreate $donationData,
+        PersonId $donorId,
+        DonationService $donationService
+    ): Donation {
         return $donationService->buildFromAPIRequest($donationData, $donorId);
     }
 
@@ -424,19 +423,6 @@ class DoctrineDonationRepository extends SalesforceProxyRepository implements Do
         }
 
         return count($donations);
-    }
-
-    public function setCampaignRepository(CampaignRepository $campaignRepository): void
-    {
-        $this->campaignRepository = $campaignRepository;
-    }
-
-    /**
-     * @param FundRepository $fundRepository
-     */
-    public function setFundRepository(FundRepository $fundRepository): void
-    {
-        $this->fundRepository = $fundRepository;
     }
 
     /**
