+++ src/Domain/DonationService.php
@@ -36,12 +36,10 @@ use Random\Randomizer;
 use Stripe\Card;
 use Stripe\Charge;
 use Stripe\Exception\ApiErrorException;
-use Stripe\Exception\CardException;
 use Stripe\Exception\InvalidRequestException;
 use Stripe\PaymentIntent;
 use Stripe\StripeObject;
 use Symfony\Component\Clock\ClockInterface;
-use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\RoutableMessageBus;
 use Symfony\Component\Notifier\ChatterInterface;
 use Symfony\Component\Notifier\Exception\TransportExceptionInterface;
@@ -92,7 +90,7 @@ class DonationService
         private RoutableMessageBus $bus,
         private DonationNotifier $donationNotifier,
         private FundRepository $fundRepository,
-) {
+    ) {
     }
 
     /**
@@ -119,6 +117,7 @@ class DonationService
         $this->rateLimiterFactory->create(key: $pspCustomerId)->consume()->ensureAccepted();
 
         try {
+            /** @psalm-suppress DeprecatedMethod */
             $donation = $this->donationRepository->buildFromApiRequest($donationData, $donorId, $this);
         } catch (\UnexpectedValueException $e) {
             $message = 'Donation Create data initial model load';
@@ -136,6 +135,8 @@ class DonationService
                 'Got campaign pull UniqueConstraintViolationException for campaign ID %s. Trying once more.',
                 $donationData->projectId->value,
             ));
+
+            /** @psalm-suppress DeprecatedMethod */
             $donation = $this->donationRepository->buildFromApiRequest($donationData, $donorId, $this);
         }
 
@@ -152,6 +153,63 @@ class DonationService
         }
 
         $this->enrollNewDonation($donation, attemptMatching: true);
+
+        return $donation;
+    }
+
+
+    /**
+     * @param DonationCreate $donationData
+     * @return Donation
+     * @throws \UnexpectedValueException if inputs invalid, including projectId being unrecognised
+     * @throws NotFoundException
+     */
+    public function buildFromAPIRequest(DonationCreate $donationData, PersonId $donorId): Donation
+    {
+
+        if (!in_array($donationData->psp, ['stripe'], true)) {
+            throw new \UnexpectedValueException(sprintf(
+                'PSP %s is invalid',
+                $donationData->psp,
+            ));
+        }
+
+        $campaign = $this->campaignRepository->findOneBy(['salesforceId' => $donationData->projectId->value]);
+
+        if (!$campaign) {
+            // Fetch data for as-yet-unknown campaigns on-demand
+            $this->logger->info("Loading unknown campaign ID {$donationData->projectId} on-demand");
+            try {
+                $campaign = $this->campaignRepository->pullNewFromSf($donationData->projectId);
+            } catch (ClientException $exception) {
+                $this->logger->error("Pull error for campaign ID {$donationData->projectId}: {$exception->getMessage()}");
+                throw new \UnexpectedValueException('Campaign does not exist');
+            }
+            $this->fundRepository->pullForCampaign($campaign);
+
+            $this->entityManager->flush();
+
+            // Because this case of campaigns being set up individually is relatively rare,
+            // it is the one place outside of `UpdateCampaigns` where we clear the whole
+            // result cache. It's currently the only user-invoked or single item place where
+            // we do so.
+            /**
+             * @psalm-suppress DeprecatedMethod
+             * @var CacheProvider $cacheDriver
+             */
+            $cacheDriver = $this->entityManager->getConfiguration()->getResultCacheImpl();
+            $cacheDriver->deleteAll();
+        }
+
+        if ($donationData->currencyCode !== $campaign->getCurrencyCode()) {
+            throw new \UnexpectedValueException(sprintf(
+                'Currency %s is invalid for campaign',
+                $donationData->currencyCode,
+            ));
+        }
+
+        $donation = Donation::fromApiModel($donationData, $campaign, $donorId);
+        $donation->deriveFees(null, null);
 
         return $donation;
     }
@@ -835,52 +893,5 @@ class DonationService
                 throw new StripeAccountIdNotSetForAccount();
             }
         }
-    }
-
-    public function buildFromAPIRequest(DonationCreate $donationData, PersonId $donorId): Donation
-    {
-
-        if (!in_array($donationData->psp, ['stripe'], true)) {
-            throw new \UnexpectedValueException(sprintf(
-                'PSP %s is invalid',
-                $donationData->psp,
-            ));
-        }
-
-        $campaign = $this->campaignRepository->findOneBy(['salesforceId' => $donationData->projectId->value]);
-
-        if (!$campaign) {
-            // Fetch data for as-yet-unknown campaigns on-demand
-            $this->logger->info("Loading unknown campaign ID {$donationData->projectId} on-demand");
-            try {
-                $campaign = $this->campaignRepository->pullNewFromSf($donationData->projectId);
-            } catch (ClientException $exception) {
-                $this->logger->error("Pull error for campaign ID {$donationData->projectId}: {$exception->getMessage()}");
-                throw new \UnexpectedValueException('Campaign does not exist');
-            }
-            $this->fundRepository->pullForCampaign($campaign);
-
-            $this->entityManager->flush();
-
-            // Because this case of campaigns being set up individually is relatively rare,
-            // it is the one place outside of `UpdateCampaigns` where we clear the whole
-            // result cache. It's currently the only user-invoked or single item place where
-            // we do so.
-            /** @var CacheProvider $cacheDriver */
-            $cacheDriver = $this->entityManager->getConfiguration()->getResultCacheImpl();
-            $cacheDriver->deleteAll();
-        }
-
-        if ($donationData->currencyCode !== $campaign->getCurrencyCode()) {
-            throw new \UnexpectedValueException(sprintf(
-                'Currency %s is invalid for campaign',
-                $donationData->currencyCode,
-            ));
-        }
-
-        $donation = Donation::fromApiModel($donationData, $campaign, $donorId);
-        $donation->deriveFees(null, null);
-
-        return $donation;
     }
 }
